+++ NEWS.md
@@ -23,6 +23,12 @@
 
  - when computing Newey-West standard-errors for time series, the bandwidth is now selected thanks to the [bwNeweyWest](https://sandwich.r-forge.r-project.org/reference/NeweyWest.html) function from the [sandwich](https://sandwich.r-forge.r-project.org/index.html) package. This function implements the method described in Newey and West 1994.
  
+## New functions
+
+ - `osize`: simple function returning a formatted object size.
+ 
+ - `n_unik`: simple but flexible function returning the number of unique elements from variables in a data set.
+ 
 ## Other
 
  - improve error messages when `subset` does not select any element.